Testing + jar file

+++ src/main/java/servicii/foto/SedintaFotoCuplu.java
@@ -17,15 +17,15 @@ public class SedintaFotoCuplu extends ServiciiFoto {
     @Override
     public String informatiiServici(){
         if(album && !studio)
-            return super.informatiiServici() + ": Sedinta foto cuplu in aer liber. Sedinta va dura " + nrMinute + " minute si va costa " + (int)pretTotal
+            return super.informatiiServici() + ": Sedinta foto cuplu in aer liber. Sedinta va dura " + (int)nrMinute + " minute si va costa " + (int)pretTotal
                     + " $. Livrarea se va face atat pe stick si veti avea si un album";
         else if(!album  && !studio)
-            return super.informatiiServici() + ": Sedinta foto cuplu in aer liber. Sedinta va dura " + nrMinute + " minute si va costa " + (int)pretTotal
+            return super.informatiiServici() + ": Sedinta foto cuplu in aer liber. Sedinta va dura " + (int)nrMinute + " minute si va costa " + (int)pretTotal
                     + " $. Livrarea se va face pe stick";
         else if(!album)
-            return super.informatiiServici() + ": Sedinta foto cuplu in studio. Sedinta va dura " + nrMinute + " minute si va costa " + (int)pretTotal
+            return super.informatiiServici() + ": Sedinta foto cuplu in studio. Sedinta va dura " + (int)nrMinute + " minute si va costa " + (int)pretTotal
                     + " $. Livrarea se va face pe stick";
-        else return super.informatiiServici() + ": Sedinta foto cuplu in studio. Sedinta va dura " + nrMinute + " minute si va costa " + (int)pretTotal
+        else return super.informatiiServici() + ": Sedinta foto cuplu in studio. Sedinta va dura " + (int)nrMinute + " minute si va costa " + (int)pretTotal
                     + " $. Livrarea se va face atat pe stick si veti avea si un album";
     }
 }
