+++ src/main/java/servicii/foto/SedintaFotoFamilie.java
@@ -20,15 +20,15 @@ public class SedintaFotoFamilie extends ServiciiFoto {
     @Override
     public String informatiiServici() {
         if(album && !studio)
-            return super.informatiiServici() + ": Sedinta foto familie in aer liber. Sedinta va dura " + nrMinute + " minute si va costa " + (int)pretTotal
+            return super.informatiiServici() + ": Sedinta foto familie in aer liber. Sedinta va dura " + (int)nrMinute + " minute si va costa " + (int)pretTotal
                     + " $. Livrarea se va face atat pe stick si veti avea si un album";
         else if(!album  && !studio)
-            return super.informatiiServici() + ": Sedinta foto familie in aer liber. Sedinta va dura " + nrMinute + " minute si va costa " + (int)pretTotal
+            return super.informatiiServici() + ": Sedinta foto familie in aer liber. Sedinta va dura " + (int)nrMinute + " minute si va costa " + (int)pretTotal
                     + " $. Livrarea se va face pe stick";
         else if(!album)
-            return super.informatiiServici() + ": Sedinta foto familie in studio. Sedinta va dura " + nrMinute + " minute si va costa " + (int)pretTotal
+            return super.informatiiServici() + ": Sedinta foto familie in studio. Sedinta va dura " + (int)nrMinute + " minute si va costa " + (int)pretTotal
                     + " $. Livrarea se va face pe stick";
-        else return super.informatiiServici() + ": Sedinta foto familie in studio. Sedinta va dura " + nrMinute + " minute si va costa " + (int)pretTotal
+        else return super.informatiiServici() + ": Sedinta foto familie in studio. Sedinta va dura " + (int)nrMinute + " minute si va costa " + (int)pretTotal
                 + " $. Livrarea se va face atat pe stick si veti avea si un album";
     }
